Users add roles, routes add permissions, some improvements

+++ routes/roles/roles.php
@@ -3,20 +3,22 @@ Route::middleware('auth')->group(function () {
     Route::group(['namespace' => 'roles'], function() {
         // views
         Route::group(['prefix' => 'roles'], function() {
-            Route::view('/', 'roles.index');
-            Route::view('/create', 'roles.create');
-            Route::view('/{role}/edit', 'roles.edit');
+            Route::view('/', 'roles.index')->middleware('permission:read-roles');
+            Route::view('/create', 'roles.create')->middleware('permission:create-roles');
+            Route::view('/{role}/edit', 'roles.edit')->middleware('permission:update-roles');
         });
 
         // api
         Route::group(['prefix' => 'api/roles'], function() {
             Route::get('/count', 'RoleController@count');
+            Route::get('/all', 'RoleController@all');
             Route::get('/getRoleModulesPermissions/{role}','RoleController@getRoleModulesPermissions');
-            Route::post('/filter', 'RoleController@filter');
-            Route::get('/{role}', 'RoleController@show');
-            Route::post('/store', 'RoleController@store');
-            Route::put('/update/{role}', 'RoleController@update');
-            Route::delete('/{user}', 'RoleController@destroy');
+            Route::post('/filter', 'RoleController@filter')->middleware('permission:read-roles');
+
+            Route::get('/{role}', 'RoleController@show')->middleware('permission:read-roles');
+            Route::post('/store', 'RoleController@store')->middleware('permission:create-roles');
+            Route::put('/update/{role}', 'RoleController@update')->middleware('permission:update-roles');
+            Route::delete('/{user}', 'RoleController@destroy')->middleware('permission:delet-roles');
         });
     });
 });
