+++ routes/users/users.php
@@ -3,19 +3,21 @@ Route::middleware('auth')->group(function () {
     Route::group(['namespace' => 'users'], function() {
         // views
         Route::group(['prefix' => 'users'], function() {
-            Route::view('/', 'users.index');
-            Route::view('/create', 'users.create');
-            Route::view('/{user}/edit', 'users.edit');
+            Route::view('/', 'users.index')->middleware('permission:read-users');
+            Route::view('/create', 'users.create')->middleware('permission:create-users');
+            Route::view('/{user}/edit', 'users.edit')->middleware('permission:update-users');
         });
 
         // api
         Route::group(['prefix' => 'api/users'], function() {
+            Route::get('/getUserRoles/{user}', 'UserController@getUserRoles');
             Route::get('/count', 'UserController@count');
-            Route::post('/filter', 'UserController@filter');
-            Route::get('/{user}', 'UserController@show');
-            Route::post('/store', 'UserController@store');
-            Route::put('/update/{user}', 'UserController@update');
-            Route::delete('/{user}', 'UserController@destroy');
+            Route::post('/filter', 'UserController@filter')->middleware('permission:read-users');
+
+            Route::get('/{user}', 'UserController@show')->middleware('permission:read-users');
+            Route::post('/store', 'UserController@store')->middleware('permission:create-users');
+            Route::put('/update/{user}', 'UserController@update')->middleware('permission:update-users');
+            Route::delete('/{user}', 'UserController@destroy')->middleware('permission:delete-users');
         });
     });
 });
